Remove Adaptation base object.
This removes the Adaptation object from which the main ChatBot
class was inheriting. Adaptation was intended to keep some of
the implementation details out of the main class to keep the
code clean and readable. However this did not work as intended.

Due to the number of emails I have received asking questions
about it, I have decided to move it back to the man ChatBot
class so that all of the initialization for the ChatBot happens
in one place.
