+++ setup.py
@@ -49,7 +49,6 @@ setup(
         'Topic :: Communications :: Chat',
         'Topic :: Internet',
         'Programming Language :: Python',
-        'Programming Language :: Python :: 2',
         'Programming Language :: Python :: 2.7',
         'Programming Language :: Python :: 3',
         'Programming Language :: Python :: 3.3',
