+++ tests/test_chatbot.py
@@ -97,3 +97,30 @@ class ChatterBotResponseTests(ChatBotTestCase):
         self.assertEqual(second_response, self.test_statement.text)
         self.assertEqual(len(statement.in_response_to), 1)
         self.assertIn("Hi", statement.in_response_to)
+
+
+class MultiAdapterTests(ChatBotTestCase):
+
+    def test_add_storage_adapter(self):
+        count_before = len(self.chatbot.storage_adapters)
+
+        self.chatbot.add_adapter(
+            "chatterbot.adapters.storage.JsonDatabaseAdapter"
+        )
+        self.assertEqual(len(self.chatbot.storage_adapters), count_before + 1)
+
+    def test_add_logic_adapter(self):
+        count_before = len(self.chatbot.logic.adapters)
+
+        self.chatbot.add_adapter(
+            "chatterbot.adapters.logic.ClosestMatchAdapter"
+        )
+        self.assertEqual(len(self.chatbot.logic.adapters), count_before + 1)
+
+    def test_add_io_adapter(self):
+        count_before = len(self.chatbot.io.adapters)
+
+        self.chatbot.add_adapter(
+            "chatterbot.adapters.io.TerminalAdapter"
+        )
+        self.assertEqual(len(self.chatbot.io.adapters), count_before + 1)
